# documentation/openapi/command-api.yaml
openapi: "3.0.3"
info:
  title: "trial API"
  description: "trial API"
  version: "1.0.0"
servers:
  - url: "https://localhost:8080"
paths:
  /api/v1/notices:
    post:
      summary: "POST api/v1/notices"
      operationId: "create"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeCreationPayload"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Void"
  /api/v1/notices/{id}:
    put:
      summary: "PUT api/v1/notices/{id}"
      operationId: "update"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeModificationPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoticeResponse"
components:
  schemas:
    NoticeCreationPayload:
      type: "object"
      properties:
        date:
          type: "string"
          format: "date-time"
        place:
          type: "string"
        pid:
          type: "string"
        casePresent:
          type: "boolean"
        type:
          type: "string"
          enum:
            - "HOSPITAL"
            - "DENTISTRY"
            - "FAMILY_DOCTOR"
            - "PRIVATE_PERSON"
        comment:
          type: "string"
    Void:
      type: "object"
      properties: { }
    NoticeModificationPayload:
      type: "object"
      properties:
        status:
          type: "string"
          enum:
            - "IN_PROCESS"
            - "APPROVED"
            - "REJECTED"
        comment:
          type: "string"
    NoticeResponse:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
        date:
          type: "string"
          format: "date-time"
        place:
          type: "string"
        pid:
          type: "string"
        casePresent:
          type: "boolean"
        type:
          type: "string"
          enum:
            - "HOSPITAL"
            - "DENTISTRY"
            - "FAMILY_DOCTOR"
            - "PRIVATE_PERSON"
        status:
          type: "string"
          enum:
            - "IN_PROCESS"
            - "APPROVED"
            - "REJECTED"
        comment:
          type: "string"
        deadlineAt:
          type: "string"
          format: "date-time"
        createdBy:
          type: "string"
        updatedBy:
          type: "string"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"